More hierarchy refactor.
