+++ python/shotgun_model/shotgun_model.py
@@ -13,7 +13,7 @@ import copy
 import os
 import sys
 import hashlib
-import weakref
+
 
 from sgtk.platform.qt import QtCore, QtGui
 
@@ -64,22 +64,15 @@ class ShotgunModel(ShotgunQueryModel):
                                  this is None then a task manager will be created as needed.
         :type bg_task_manager: :class:`~task_manager.BackgroundTaskManager`
         """
-        super(ShotgunModel, self).__init__(parent, bg_task_manager)
+        super(ShotgunModel, self).__init__(parent, bg_load_thumbs, bg_task_manager)
 
         # default value so that __repr__ can be used before load_data
         self.__entity_type = None
 
         self.__schema_generation = schema_generation
 
-        # is the model set up with a query?
-        self.__has_query = False
-
         # keep track of info for thumbnail download/load
         self.__download_thumbs = download_thumbs
-        self.__bg_load_thumbs = bg_load_thumbs
-        self.__thumb_map = {}
-
-        self.__current_work_id = None
 
     def __repr__(self):
         """
@@ -100,15 +93,6 @@ class ShotgunModel(ShotgunQueryModel):
         # note that all of the ids may not be loaded into the actual model
         # yet so we have to query the data handler for this.
         return self._data_handler.get_entity_ids()
-
-    def destroy(self):
-        """
-        Call this method prior to destroying this object.
-        This will ensure all worker threads etc are stopped.
-        """
-        self.__current_work_id = None
-        self.__thumb_map = {}
-        super(ShotgunModel, self).destroy()
 
     def item_from_entity(self, entity_type, entity_id):
         """
@@ -132,55 +116,7 @@ class ShotgunModel(ShotgunQueryModel):
             self._log_debug("...entity id is not part of the data set")
             return None
 
-        # this node is loaded in the query cached by the data handler
-        # but may not exist in the model yet - because of deferred loading.
-
-        # first see if we have it in the tree
-        item = self._get_item_by_unique_id(uid)
-
-        if not item:
-
-            data_item = self._data_handler.get_data_item_from_uid(uid)
-
-            # item was not part of the model. Attempt to load its parents until it is visible.
-            self._log_debug("Item %s does not exist in the tree - will expand tree." % data_item)
-
-            # now get a list of all parents and recurse back down towards
-            # the node we want to load. If at any level, the data has not
-            # yet been loaded, we expand that level.
-            hierarchy_bottom_up = []
-            node = data_item
-            while node:
-                hierarchy_bottom_up.append(node)
-                node = node.parent
-
-            # reverse the list to get the top-down hierarchy
-            hierarchy_top_down = hierarchy_bottom_up[::-1]
-
-            self._log_debug("Resolved top-down hierarchy to be %s" % hierarchy_top_down)
-
-            for data_item in hierarchy_top_down:
-                # see if we have this item in the tree
-                item = self._get_item_by_unique_id(data_item.unique_id)
-                if not item:
-                    self._log_debug(
-                        "Data item %s does not exist in model - fetching parent's children..." % data_item
-                    )
-                    # this parent does not yet exist in the tree
-                    # find the parent and kick it to expand it
-
-                    # assume that the top level is always loaded in tree
-                    # so that it's always safe to do data_item.parent.uid here
-                    parent_item = self._get_item_by_unique_id(data_item.parent.unique_id)
-                    # get model index
-                    parent_model_index = parent_item.index()
-                    # kick it
-                    self.fetchMore(parent_model_index)
-
-            # now try again
-            item = self._get_item_by_unique_id(uid)
-
-        return item
+        return self._ensure_item_loaded(uid)
 
     def index_from_entity(self, entity_type, entity_id):
         """
@@ -250,33 +186,6 @@ class ShotgunModel(ShotgunQueryModel):
         """
         # column is one greater than the index because of the default initial column
         return [{"column_idx": i + 1, "field": field} for (i, field) in enumerate(self.__column_fields)]
-
-    def hard_refresh(self):
-        """
-        Clears any caches on disk, then refreshes the data.
-        """
-        if not self.__has_query:
-            # no query in this model yet
-            return
-
-        super(ShotgunModel, self).hard_refresh()
-
-    ########################################################################################
-    # methods overridden from the base class.
-
-    def clear(self):
-        """
-        Removes all items (including header items) from the model and
-        sets the number of rows and columns to zero.
-        """
-
-        # clear thumbnail download lookup so we don't process any more results:
-        self.__thumb_map = {}
-
-        # we are not looking for any data from the async processor
-        self.__current_work_id = None
-
-        super(ShotgunModel, self).clear()
 
     ########################################################################################
     # protected methods not meant to be subclassed but meant to be called by subclasses
@@ -365,7 +274,6 @@ class ShotgunModel(ShotgunQueryModel):
         # clear out old data
         self.clear()
 
-        self.__has_query = True
         self.__entity_type = entity_type
         self.__filters = filters
         self.__fields = fields
@@ -373,7 +281,7 @@ class ShotgunModel(ShotgunQueryModel):
         self.__hierarchy = hierarchy
         self.__column_fields = columns or []
         self.__editable_fields = editable_columns or []
-        self.__limit = limit or 0 # 0 means get all matches
+        self.__limit = limit or 0  # 0 means get all matches
         self.__additional_filter_presets = additional_filter_presets
 
         # make sure `editable_fields` is a subset of `column_fields`
@@ -395,7 +303,19 @@ class ShotgunModel(ShotgunQueryModel):
         self._log_debug("Filter Presets: %s" % self.__additional_filter_presets)
 
         # get the cache path based on these new data query parameters
-        self._data_handler = ShotgunFindDataHandler(self.__compute_cache_path(seed), self)
+        self._data_handler = ShotgunFindDataHandler(
+            self.__entity_type,
+            self.__filters,
+            self.__order,
+            self.__hierarchy,
+            self.__fields,
+            self.__column_fields,
+            self.__download_thumbs,
+            self.__limit,
+            self.__additional_filter_presets,
+            self.__compute_cache_path(seed),
+            self
+        )
         # load up from disk
         self._log_debug("Loading data from cache file into memory...")
         self._data_handler.load_cache()
@@ -424,119 +344,6 @@ class ShotgunModel(ShotgunQueryModel):
         # return true if cache is loaded false if not
         return nodes_generated > 0
 
-    def _refresh_data(self):
-        """
-        Rebuilds the data in the model to ensure it is up to date.
-        This call is asynchronous and will return instantly.
-        The update will be applied whenever the data from Shotgun is returned.
-
-        If the model is empty (no cached data) no data will be shown at first
-        while the model fetches data from Shotgun.
-
-        As soon as a local cache exists, data is shown straight away and the
-        shotgun update happens silently in the background.
-
-        If data has been added, this will be injected into the existing structure.
-        In this case, the rest of the model is intact, meaning that also selections
-        and other view related states are unaffected.
-
-        If data has been modified or deleted, a full rebuild is issued, meaning that
-        all existing items from the model are removed. This does affect view related
-        states such as selection.
-        """
-        if not self._sg_data_retriever:
-            raise sgtk.TankError("Data retriever is not available!")
-
-        # Stop any queued work that hasn't completed yet.  Note that we intentionally only stop the
-        # find query and not the thumbnail cache/download.  This is because the thumbnails returned
-        # are likely to still be valid for the current data in the model and if they are stopped then
-        # the pattern 'create model->load cached->refresh from sg' would result in empty icons being
-        # presented to the user until the shotgun query has completed!
-        #
-        # This may result in unnecessary thumbnail downloads from Shotgun but in all likelihood, the
-        # thumbnails are going to be the same before and after the refresh and any additional overhead
-        # should be weighed against a cleaner user experience
-        if self.__current_work_id is not None:
-            self._sg_data_retriever.stop_work(self.__current_work_id)
-            self.__current_work_id = None
-
-        # emit that the data is refreshing.
-        self.data_refreshing.emit()
-
-        if self.__filters is None:
-            # filters is None indicates that no data is desired.
-            # do not issue the sg request but pass straight to the callback
-            self.__on_sg_data_arrived([])
-        else:
-            # get data from shotgun - list/set cast to ensure unique fields
-            fields = self.__hierarchy + self.__fields + self.__column_fields
-            if self.__download_thumbs:
-                fields = fields + ["image"]
-            fields = list(set(fields))
-
-            find_kwargs = dict(
-                limit=self.__limit,
-            )
-
-            # We only want to include the filter presets kwarg if it was explicitly asked
-            # for. The reason for this is that it's a Shotgun 7.0 feature server side, and
-            # we don't want to break backwards compatibility with older versions of Shotgun.
-            if self.__additional_filter_presets:
-                find_kwargs["additional_filter_presets"] = self.__additional_filter_presets
-
-            self.__current_work_id = self._sg_data_retriever.execute_find(
-                self.__entity_type,
-                self.__filters,
-                fields,
-                self.__order,
-                **find_kwargs
-            )
-
-    def _request_thumbnail_download(self, item, field, url, entity_type, entity_id):
-        """
-        Request that a thumbnail is downloaded for an item. If a thumbnail is successfully
-        retrieved, either from disk (cached) or via shotgun, the method _populate_thumbnail()
-        will be called. If you want to control exactly how your shotgun thumbnail is
-        to appear in the UI, you can subclass this method. For example, you can subclass
-        this method and perform image composition prior to the image being added to
-        the item object.
-
-        .. note:: This is an advanced method which you can use if you want to load thumbnail
-            data other than the standard 'image' field. If that's what you need, simply make
-            sure that you set the download_thumbs parameter to true when you create the model
-            and standard thumbnails will be automatically downloaded. This method is either used
-            for linked thumb fields or if you want to download thumbnails for external model data
-            that doesn't come from Shotgun.
-
-        :param item: :class:`~PySide.QtGui.QStandardItem` which belongs to this model
-        :param field: Shotgun field where the thumbnail is stored. This is typically ``image`` but
-                      can also for example be ``sg_sequence.Sequence.image``.
-        :param url: thumbnail url
-        :param entity_type: Shotgun entity type
-        :param entity_id: Shotgun entity id
-        """
-        if url is None:
-            # nothing to download. bad input. gracefully ignore this request.
-            return
-
-        if not self._sg_data_retriever:
-            raise sgtk.ShotgunModelError("Data retriever is not available!")
-
-        uid = self._sg_data_retriever.request_thumbnail(
-            url,
-            entity_type,
-            entity_id,
-            field,
-            self.__bg_load_thumbs
-        )
-
-        # keep tabs of this and call out later - note that we use a weakref to allow
-        # the model item to be gc'd if it's removed from the model before the thumb
-        # request completes.
-        self.__thumb_map[uid] = {
-            "item_ref": weakref.ref(item),
-            "field": field
-        }
 
     ########################################################################################
     # methods to be implemented by subclasses
@@ -555,13 +362,29 @@ class ShotgunModel(ShotgunQueryModel):
         :param item: The item that was just created.
         :type item: :class:`~PySide.QtGui.QStandardItem`
         """
-
         # as per docs, call the base implementation
         super(ShotgunModel, self)._item_created(item)
 
-        # request thumb
+        # request thumbnail for this item
         if self.__download_thumbs:
-            self.__process_thumbnail_for_item(item)
+            sg_data = item.data(self.SG_DATA_ROLE)
+
+            if sg_data:
+                for field in sg_data.keys():
+                    # note: we check for all fields containing "image"
+                    # so that we'll catch a field such as 'sg_sequence.Sequence.image'
+                    # as well as a straight 'image' field
+                    if "image" in field and sg_data[field] is not None:
+                        # we have a thumb we are supposed to download!
+                        # get the thumbnail - store the unique id we get back from
+                        # the data retrieve in a dict for fast lookup later
+                        self._request_thumbnail_download(
+                            item,
+                            field,
+                            sg_data[field],
+                            sg_data.get("type"),
+                            sg_data.get("id")
+                        )
 
     def _set_tooltip(self, item, sg_item):
         """
@@ -629,59 +452,6 @@ class ShotgunModel(ShotgunQueryModel):
                     item.text()
                 )
             )
-
-    def _populate_thumbnail(self, item, field, path):
-        """
-        Called whenever the real thumbnail for an item exists on disk. The following
-        execution sequence typically happens:
-
-        - :class:`~PySide.QtGui.QStandardItem` is created, either through a cache load from disk or
-          from a payload coming from the Shotgun API.
-        - After the item has been set up with its associated Shotgun data,
-          :meth:`_populate_default_thumbnail()` is called, allowing client code to set
-          up a default thumbnail that will be shown while potential real thumbnail
-          data is being loaded.
-        - The model will now start looking for the real thumbail.
-        - If the thumbnail is already cached on disk, :meth:`_populate_thumbnail()` is called very soon.
-        - If there isn't a thumbnail associated, :meth:`_populate_thumbnail()` will not be called.
-        - If there isn't a thumbnail cached, the model will asynchronously download
-          the thumbnail from Shotgun and then (after some time) call :meth:`_populate_thumbnail()`.
-
-        This method will be called for standard thumbnails if the model has been
-        instantiated with the download_thumbs flag set to be true. It will be called for
-        items which are associated with shotgun entities (in a tree data layout, this is typically
-        leaf nodes). It will also be called once the data requested via _request_thumbnail_download()
-        arrives.
-
-        This method makes it possible to control how the thumbnail is applied and associated
-        with the item. The default implementation will simply set the thumbnail to be icon
-        of the item, but this can be altered by subclassing this method.
-
-        :param item: :class:`~PySide.QtGui.QStandardItem` which is associated with the given thumbnail
-        :param field: The Shotgun field which the thumbnail is associated with.
-        :param path: A path on disk to the thumbnail. This is a file in jpeg format.
-        """
-        # the default implementation sets the icon
-        thumb = QtGui.QPixmap(path)
-        item.setIcon(thumb)
-
-    def _populate_thumbnail_image(self, item, field, image, path):
-        """
-        Similar to :meth:`_populate_thumbnail()` but this method is called instead
-        when the bg_load_thumbs parameter has been set to true. In this case, no
-        loading of thumbnail data from disk is necessary - this has already been
-        carried out async and is passed in the form of a QImage object.
-
-        For further details, see :meth:`_populate_thumbnail()`
-
-        :param item: :class:`~PySide.QtGui.QStandardItem` which is associated with the given thumbnail
-        :param field: The Shotgun field which the thumbnail is associated with.
-        :param image: QImage object with the thumbnail loaded
-        :param path: A path on disk to the thumbnail. This is a file in jpeg format.
-        """
-        # the default implementation sets the icon
-        thumb = QtGui.QPixmap.fromImage(image)
-        item.setIcon(thumb)
 
     def _get_additional_columns(self, primary_item, is_leaf, columns):
         """
@@ -758,76 +528,6 @@ class ShotgunModel(ShotgunQueryModel):
         row.extend(self._get_additional_columns(item, is_leaf, self.__column_fields))
         return row
 
-    def _on_data_retriever_work_failure(self, uid, msg):
-        """
-        Asynchronous callback - the data retriever failed to do some work
-
-        :param uid: The unique id of the work that failed
-        :param msg: The error message returned for the failure
-        """
-        uid = sanitize_qt(uid) # qstring on pyqt, str on pyside
-        msg = sanitize_qt(msg)
-
-        if self.__current_work_id != uid:
-            # not our job. ignore
-            self._log_debug("Retrieved error from data worker: %s" % msg)
-            return
-        self.__current_work_id = None
-
-        full_msg = "Error retrieving data from Shotgun: %s" % msg
-        self.data_refresh_fail.emit(full_msg)
-        self._log_warning(full_msg)
-
-    def _on_data_retriever_work_completed(self, uid, request_type, data):
-        """
-        Signaled whenever the data retriever completes some work.
-        This method will dispatch the work to different methods
-        depending on what async task has completed.
-
-        :param uid:             The unique id of the work that completed
-        :param request_type:    Type of work completed
-        :param data:            Result of the work
-        """
-        uid = sanitize_qt(uid) # qstring on pyqt, str on pyside
-        data = sanitize_qt(data)
-
-        self._log_debug("Received worker payload of type %s" % request_type)
-
-        if self.__current_work_id == uid:
-            # our data has arrived from sg!
-            # process the data
-            self.__current_work_id = None
-            sg_data = data["sg"]
-            self.__on_sg_data_arrived(sg_data)
-
-        elif uid in self.__thumb_map:
-            # a thumbnail is now present on disk!
-            thumb_info = self.__thumb_map[uid]
-            del self.__thumb_map[uid]
-
-            thumbnail_path = data["thumb_path"]
-            thumbnail = data["image"]
-
-            # if the requested thumbnail has since dissapeared on the server,
-            # path and image will be None. In this case, skip processing
-            if thumbnail_path:
-                # get the model item from the weakref we stored in the thumb info:
-                item = thumb_info["item_ref"]()
-                if not item:
-                    # the model item no longer exists so we can ignore this result!
-                    return
-                sg_field = thumb_info["field"]
-
-                # call our deriving class implementation
-                if self.__bg_load_thumbs:
-                    # worker thread already loaded the thumbnail in as a QImage.
-                    # call a separate method.
-                    self._populate_thumbnail_image(item, sg_field, thumbnail, thumbnail_path)
-                else:
-                    # worker thread only ensured that the image exists
-                    # call method to populate it
-                    self._populate_thumbnail(item, sg_field, thumbnail_path)
-
     def _create_item(self, parent, data_item):
         """
         Creates a model item for the tree given data out of the data store
@@ -842,7 +542,7 @@ class ShotgunModel(ShotgunQueryModel):
         item = ShotgunStandardItem()
         item.setEditable(data_item.field in self.__editable_fields)
 
-        self.__update_item(item, data_item)
+        self._update_item(item, data_item)
 
         # run the finalizer
         self._finalize_item(item)
@@ -855,88 +555,7 @@ class ShotgunModel(ShotgunQueryModel):
 
         return item
 
-    ########################################################################################
-    # private methods
-
-    def __save_data_async(self, sg):
-        """
-        Asynchronous callback to perform a cache save in the background.
-
-        :param :class:`Shotgun` sg: Shotgun API instance
-        """
-        self._log_debug("Begin asynchronously saving cache to disk")
-        self._data_handler.save_cache()
-        self._log_debug("Asynchronous cache save complete.")
-
-    def __on_sg_data_arrived(self, sg_data):
-        """
-        Handle asynchronous shotgun data arriving after a find request.
-
-        :param list sg_data: Shotgun data payload.
-        """
-        self._log_debug("--> Shotgun data arrived. (%s records)" % len(sg_data))
-
-        # pre-process data
-        sg_data = self._before_data_processing(sg_data)
-
-        # push shotgun data into our data handler which will figure out
-        # if there are any changes
-        self._log_debug("Updating data model with new shotgun data...")
-        modified_items = self._data_handler.update_find_data(sg_data, self.__hierarchy)
-
-        self._log_debug("Shotgun data contained %d modifications" % len(modified_items))
-
-        if len(modified_items) > 0:
-            # save cache changes to disk in the background
-            self._sg_data_retriever.execute_method(self.__save_data_async)
-
-        root = self.invisibleRootItem()
-        if root.rowCount() == 0:
-            # an empty tree - in this case perform a full insert, not a diff
-            self._log_debug("Model was empty - doing a full load pass...")
-            self._data_handler.generate_child_nodes(None, root, self._create_item)
-
-        else:
-            # the tree was already loaded. Perform diffs instead.
-            # update tree based on modifications
-            for item in modified_items:
-                data_item = item["data"]
-
-                self._log_debug("Processing change %s" % item)
-
-                if item["mode"] == self._data_handler.ADDED:
-                    # look for the parent of this item
-                    parent_data_item = data_item.parent
-                    # see if this exists in the tree
-                    parent_model_item = self._get_item_by_unique_id(parent_data_item.unique_id)
-                    if parent_model_item:
-                        # the parent exists in the view. So add the child
-                        # note: becuase of lazy loading, parent may not always exist.
-                        self._log_debug("Creating new model item for %s" % data_item)
-                        self._create_item(parent_model_item, data_item)
-
-                elif item["mode"] == self._data_handler.DELETED:
-                    # see if the node exists in the tree, in that case delete it.
-                    # we check if it exists in the model because it may not have been
-                    # loaded in yet by the deferred loader
-                    model_item = self._get_item_by_unique_id(data_item.unique_id)
-                    if model_item:
-                        self._log_debug("Deleting model subtree %s" % model_item)
-                        self._delete_item(model_item)
-
-                elif item["mode"] == self._data_handler.UPDATED:
-                    # see if the node exists in the tree, in that case update it with new info
-                    # we check if it exists in the model because it may not have been
-                    # loaded in yet by the deferred loader
-                    model_item = self._get_item_by_unique_id(data_item.unique_id)
-                    if model_item:
-                        self._log_debug("Updating model item %s" % model_item)
-                        self.__update_item(model_item, data_item)
-
-        # and emit completion signal
-        self.data_refreshed.emit(modified_items > 0)
-
-    def __update_item(self, item, data_item):
+    def _update_item(self, item, data_item):
         """
         Updates a model item with the given data
 
@@ -989,6 +608,9 @@ class ShotgunModel(ShotgunQueryModel):
 
         self._set_tooltip(item, data_item.shotgun_data)
 
+    ########################################################################################
+    # private methods
+
     def __compute_cache_path(self, cache_seed=None):
         """
         Calculates and returns a cache path to use for this instance's query.
@@ -1033,7 +655,6 @@ class ShotgunModel(ShotgunQueryModel):
         # now hash up the rest of the parameters and make that the filename
         params_hash = hashlib.md5()
         params_hash.update(str(self.__schema_generation))
-        params_hash.update(str(self.__bg_load_thumbs))
         params_hash.update(str(self.__fields))
         params_hash.update(str(self.__order))
         params_hash.update(str(self.__hierarchy))
@@ -1070,28 +691,6 @@ class ShotgunModel(ShotgunQueryModel):
             )
 
         return data_cache_path
-
-    def __process_thumbnail_for_item(self, item):
-        """
-        Schedule a thumb download for an item
-        """
-        sg_data = item.data(self.SG_DATA_ROLE)
-
-        if sg_data is None:
-            return
-
-        for field in sg_data.keys():
-            if "image" in field and sg_data[field] is not None:
-                # we have a thumb we are supposed to download!
-                # get the thumbnail - store the unique id we get back from
-                # the data retrieve in a dict for fast lookup later
-                self._request_thumbnail_download(
-                    item,
-                    field,
-                    sg_data[field],
-                    sg_data.get("type"),
-                    sg_data.get("id")
-                )
 
     def __generate_display_name(self, field, sg_data):
         """
